app/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
openai>=1.0.0
pymupdf
pdfplumber
pdf2image
pytesseract
pillow
langchain
langchain-text-splitters
faiss-cpu
numpy
tqdm
requests
sentence-transformers
pymongo
